[Backend] Renamed entity to entities and migration to migrations
