+++ backend/src/index.ts
@@ -1,14 +1,14 @@
 import { appDataSource } from './config/data-source';
-import { User } from './entity/User';
+import { User } from './entities/User';
 
 appDataSource
   .initialize()
   .then(async () => {
     console.log('Inserting a new user into the database...');
     const user = new User();
-    user.firstName = 'Timber';
-    user.lastName = 'Saw';
-    user.age = 25;
+    user.email = 'Timber';
+    user.password = 'Saw';
+
     await appDataSource.manager.save(user);
     console.log('Saved a new user with id: ' + user.id);
